Further improvements to TestChunkInputStream
